dbaops/generators/data_generator/requirements.txt
psycopg[binary]>=3.2.0
pymysql>=1.1.0
confluent-kafka>=2.5.0
aws-msk-iam-sasl-signer-python>=1.0.0
boto3>=1.34.0
